Added Deposit functionality
Added screens for Withdraw and Transfer
Added account number generation
Modified UpdateCustomer method to sync the Id of new accounts back to
the detached entity

+++ Banking/Banking/Domain/Services/BankingOperationsEngine/IAccountOperationsManager.cs
@@ -4,7 +4,7 @@ namespace Banking.Domain.Services.BankingOperationsEngine
 {
     public interface IAccountOperationsManager
     {
-        void Deposit(IAccount account, decimal amount);
+        void Deposit(IAccount account, double amount);
 
         void Withdraw(IAccount account, decimal amount);
